+++ CHANGELOG.md
@@ -6,15 +6,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-rc.2] - 2025-12-20
+
+### Added
+
+- Add `schemaVersion` to recipe schema
+
+### Changed
+
+- Make `links` field optional, and thus `undefined` in the recipe object, unless link parsing is enabled
+- Move schema transform & refinement validations into a helper function `applyRecipeValidations`
+
 ## [1.0.0-rc.1] - 2025-12-20
 
 ### Added
 
-- chore: tsdown configuration file
+- Add `tsdown` configuration file
 
 ### Fixed
 
-- fix: main/module/type entriess in package.json; add exports field
+- Fix main/module/type entriess in package.json; add exports field
 
 ## [1.0.0-rc.0] - 2025-12-20